pyproject.toml
[tool.poetry]
name = "github-actions-example"
version = "0.1.0"
description = ""
authors = ["Fernando Martínez González <frndmartinezglez@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
pytest-xdist = "^2.1.0"
flake8 = "^3.8.4"
pytest-cov = "^2.10.1"
hypothesis = "^5.41.4"

[tool.commitizen]
version = "0.3.0"
version_files = [
    "primes/__init__.py",
]
tag_format = "v$version"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

